[REFACTOR] Wardrobe code splitting
