+++ pandora-client-web/src/components/wardrobe/itemDetail/_wardrobeItemDetail.tsx
@@ -1,0 +1,147 @@
+import {
+	ItemPath,
+} from 'pandora-common';
+import React, { ReactElement, useCallback, useEffect } from 'react';
+import { FieldsetToggle } from '../../common/fieldsetToggle';
+import { Column, Row } from '../../common/container/container';
+import { ItemModuleLockSlot } from 'pandora-common/dist/assets/modules/lockSlot';
+import { SplitContainerPath } from 'pandora-common/dist/assets/appearanceHelpers';
+import { WardrobeFocus } from '../wardrobeTypes';
+import { useWardrobeContext } from '../wardrobeContext';
+import { useWardrobeTargetItem } from '../wardrobeUtils';
+import { WardrobeActionButton } from '../wardrobeComponents';
+import { WardrobeItemColorization } from './wardrobeItemColor';
+import { WardrobeModuleConfig } from '../modules/_wardrobeModules';
+import { WardrobeRoomDeviceDeployment, WardrobeRoomDeviceSlots, WardrobeRoomDeviceWearable } from './wardrobeItemRoomDevice';
+
+export function WardrobeItemConfigMenu({
+	item,
+	setFocus,
+}: {
+	item: ItemPath;
+	setFocus: (newFocus: WardrobeFocus) => void;
+}): ReactElement {
+	const { targetSelector, target } = useWardrobeContext();
+	const wornItem = useWardrobeTargetItem(target, item);
+
+	const containerPath = SplitContainerPath(item.container);
+	const containerItem = useWardrobeTargetItem(target, containerPath?.itemPath);
+	const containerModule = containerPath != null ? containerItem?.modules.get(containerPath.module) : undefined;
+	const singleItemContainer = containerModule != null && containerModule instanceof ItemModuleLockSlot;
+	const isRoomInventory = target.type === 'room' && item.container.length === 0;
+
+	const close = useCallback(() => {
+		setFocus({
+			container: item.container,
+			itemId: null,
+		});
+	}, [item, setFocus]);
+
+	useEffect(() => {
+		if (!wornItem) {
+			close();
+		}
+	}, [wornItem, close]);
+
+	if (!wornItem) {
+		return (
+			<div className='inventoryView'>
+				<div className='toolbar'>
+					<span>Editing item: [ ERROR: ITEM NOT FOUND ]</span>
+					<button className='modeButton' onClick={ close }>✖️</button>
+				</div>
+			</div>
+		);
+	}
+
+	return (
+		<div className='inventoryView'>
+			<div className='toolbar'>
+				<span>Editing item: { wornItem.asset.definition.name }</span>
+				{ !singleItemContainer && <button className='modeButton' onClick={ close }>✖️</button> }
+			</div>
+			<Column padding='medium' overflowX='hidden' overflowY='auto'>
+				<Row padding='medium' wrap>
+					{
+						singleItemContainer ? null : (
+							<>
+								<WardrobeActionButton action={ {
+									type: 'move',
+									target: targetSelector,
+									item,
+									shift: 1,
+								} }>
+									▼ Wear on top
+								</WardrobeActionButton>
+								<WardrobeActionButton action={ {
+									type: 'move',
+									target: targetSelector,
+									item,
+									shift: -1,
+								} }>
+									▲ Wear under
+								</WardrobeActionButton>
+							</>
+						)
+					}
+					<WardrobeActionButton
+						action={ {
+							type: 'delete',
+							target: targetSelector,
+							item,
+						} }
+						onExecute={ close }
+					>
+						➖ Remove and delete
+					</WardrobeActionButton>
+					{
+						!isRoomInventory ? (
+							<WardrobeActionButton
+								action={ {
+									type: 'transfer',
+									source: targetSelector,
+									item,
+									target: { type: 'roomInventory' },
+									container: [],
+								} }
+								onExecute={ close }
+							>
+								<span>
+									<u>▽</u> Store in room
+								</span>
+							</WardrobeActionButton>
+						) : null
+					}
+				</Row>
+				{
+					(wornItem.isType('personal') || wornItem.isType('roomDevice')) ? (
+						<WardrobeItemColorization wornItem={ wornItem } item={ item } />
+					) : null
+				}
+				{
+					wornItem.isType('roomDevice') ? (
+						<WardrobeRoomDeviceDeployment roomDevice={ wornItem } item={ item } />
+					) : null
+				}
+				{
+					wornItem.isType('roomDevice') ? (
+						<WardrobeRoomDeviceSlots roomDevice={ wornItem } item={ item } />
+					) : null
+				}
+				{
+					wornItem.isType('roomDeviceWearablePart') ? (
+						<WardrobeRoomDeviceWearable roomDeviceWearable={ wornItem } item={ item } />
+					) : null
+				}
+				{
+					Array.from(wornItem.modules.entries())
+						.map(([moduleName, m]) => (
+							<FieldsetToggle legend={ `Module: ${m.config.name}` } key={ moduleName }>
+								<WardrobeModuleConfig item={ item } moduleName={ moduleName } m={ m } setFocus={ setFocus } />
+							</FieldsetToggle>
+						))
+				}
+			</Column>
+		</div>
+	);
+}
