+++ pandora-client-web/src/editor/components/bones/bones.tsx
@@ -7,7 +7,8 @@ import { FieldsetToggle } from '../../../components/common/fieldsetToggle';
 import { Scrollbar } from '../../../components/common/scrollbar/scrollbar';
 import { ModalDialog } from '../../../components/dialog/dialog';
 import { ContextHelpButton } from '../../../components/help/contextHelpButton';
-import { BoneRowElement, WardrobeArmPoses, WardrobeExpressionGui, WardrobePoseCategories } from '../../../components/wardrobe/wardrobe';
+import { BoneRowElement, WardrobeArmPoses, WardrobePoseCategories } from '../../../components/wardrobe/views/wardrobePoseView';
+import { WardrobeExpressionGui } from '../../../components/wardrobe/views/wardrobeExpressionsView';
 import { useObservable } from '../../../observable';
 import { useEditor } from '../../editorContextProvider';
 import { EditorCharacter, useEditorCharacterState } from '../../graphics/character/appearanceEditor';
